requirements.txt
Django>=4.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
reportlab>=4.0.0
Pillow>=10.1.0
lxml>=4.9.3
python-dotenv>=1.0.0
webdriver-manager>=4.0.1